split out categories, and fixed search
